chore(cli): split man and markdown generators into separate xtasks
- introduces new `genmd` xtask for generating Markdown CLI docs
- updates `genman` xtask to generate only manpages
- removes unused `const_format` and related crates
- simplifies genman logic by dropping Markdown generation
- adds `genmd` cargo alias for convenience
- adjusts workspace dependencies and cleanup Cargo.lock
