+++ xtasks/genmd/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "genman"
+name = "genmd"
 version.workspace = true
 edition.workspace = true
 repository.workspace = true
@@ -9,6 +9,6 @@ license.workspace = true
 af.workspace = true
 anyhow.workspace = true
 clap.workspace = true
-clap_mangen = "0.2.26"
+clap-markdown = "0.1.4"
 
 [build-dependencies]
